Consolidate utils into session.
